Fix null reference exception when calculating effective date
This commit fixes a recent null reference exception raised in production
when attempting to perform a means assessment with a new work reason of
_previous assessment was incorrect_.

In the case of a benefits-bypass where the application is initially
passported before the applicant employment status is changed, a caseworker
is unable to create the means assessment as a _first means assessment_.
Should they choose _previous assessment was incorrect_ as the new work reason,
the effective date from the _CalculateContributionDTO_ is retrieved and
`toString()`'d, however because the DTO is null this throws an exception.

The migrated stored procedure `calc_contribs` (in the `contribution` package)
performs an almost-identical check as is done in the _MaatCalculateContributionService_
`getEffectiveDateByNewWorkReason` method, except that if either the new or the previous
monthly contributions are null it instead defaults back to using the latest of the
assessment committal date and either the passport assessment date, full means assessment
date or the initial means assessment date (the first of these which is not null).

By contrast, the monthly contributions in the _CalculateContributionDTO_ were always
set to a non-null value by the builder, which is a difference in logic compared to the
stored procedure and which meant that in the benefits-bypass scenerio described above,
the _CalculateContributionDTO_ would always be accessed regardless of whether there
were any contributions.

This commit fixes this issue by removing the default assignment from the builder
and re-introduces null checks against the monthly contributions both in the
`getEffectiveDateByNewWorkReason` method and the `shouldCreateContributions` methods.