Temporarily allow bypassing of Snyk in CircleCI pipeline
This commit temporarily allows the bypassing of Snyk in the CircleCI pipeline.
The reason for this commit is to address a bug in production, which is currently
blocked from being fixed because of an unrelated security vulnerability in
Spring Boot 3.3.5. Given that the Spring Boot upgrade requires other dependencies
to be updated, this is a change which should be tested more in isolation and
therefore should be done separately of the bug fix.