Fix getEffectiveDateByNewWorkReason to use correct monthly contributions
This commit fixes the `getEffectiveDateByNewWorkReason` method of the
_MaatCalculateContributionService_ to use the correct value for the
_new_ monthly contributions, as was the case in the now-migrated
`calc_contribs` stored procedure which this code originates from.

Previously, the `monthlyContributions` value being passed into this
method was set to the contributions cap which was incorrect and was
introduced in [PR #89](https://github.com/ministryofjustice/laa-crown-court-contribution/pull/89).

This commit also refactors some variable naming and makes the tests
around `getEffectiveDateByNewWorkReason` easier to understand.