Use active contribution when checking contribution equivalence
This commit fixes the `shouldCreateContribution` method of the
_CompareContributionService_ to pass the _active_ contribution
to be checked against the newly calculated contribution amount
when determining whether to create a new contribution record
for an application.

This check used to be in place but was changed in PR
https://github.com/ministryofjustice/laa-crown-court-contribution/pull/185
as part of general refactoring and was not caught as part of
that work as there were no tests covering this (all tests used
a single contribution, which meant that the issue was not raised).
Tests have been added as part of this commit to help prevent any
further regression.

This issue came to light after creating means assessments against
an application in MAAT resulted in contributions being created,
even when the contributions were identical; the reason for this
was because the first contribution was always being used for the
comparison, even when it had been replaced.